Remove the option for disabling NREUM on AMP pages
There's no context where you would want to include New Relic scripts in
AMP pages, so it doesn't make sense to expose this as an option. This
removes the option, and just hooks `disable_nr_autorum()` onto the
`pre_amp_render_post` action, so that any pages which are served through
the AMP plugin will not include the browser monitoring instrument
script.